Imported OpenTracker-current.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@15764 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/tracker/QueryPoseView.cpp
@@ -454,7 +454,7 @@ QueryEntryListCollection::QueryEntryListCollection(Model *model, BHandler *targe
 	if (model->Node()->GetAttrInfo(kAttrQueryVolume, &info) == B_OK) {
 		char *buffer = NULL;
 
-		if ((buffer = (char *)malloc(info.size)) != NULL
+		if ((buffer = (char *)malloc((size_t)info.size)) != NULL
 			&& model->Node()->ReadAttr(kAttrQueryVolume, B_MESSAGE_TYPE, 0, buffer,
 					(size_t)info.size) == info.size) {
